SCHED-515: Very unsatisfying resolution of circular imports.
