+++ scheduler/core/components/changemonitor/__init__.py
@@ -1,5 +1,5 @@
 # Copyright (c) 2016-2024 Association of Universities for Research in Astronomy, Inc. (AURA)
 # For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-from change_monitor import *
-from time_coordinate_record import *
+from .change_monitor import *
+from .time_coordinate_record import *
