+++ scheduler/scripts/run_main.py
@@ -1,20 +1,26 @@
 # Copyright (c) 2016-2024 Association of Universities for Research in Astronomy, Inc. (AURA)
 # For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-from datetime import datetime
+from datetime import datetime, timedelta
+from typing import Dict, FrozenSet, Optional
 
+import numpy as np
+from astropy.time import Time
+from lucupy.minimodel import NightIndex, TimeslotIndex
 from lucupy.minimodel.constraints import CloudCover, ImageQuality, VariantChange
-from lucupy.minimodel.semester import SemesterHalf
+from lucupy.minimodel.semester import Semester, SemesterHalf
+from lucupy.minimodel.site import ALL_SITES, Site
 from lucupy.observatory.abstract import ObservatoryProperties
 from lucupy.observatory.gemini import GeminiProperties
 
 from scheduler.core.builder.blueprint import CollectorBlueprint, OptimizerBlueprint
 from scheduler.core.builder.validationbuilder import ValidationBuilder
-from scheduler.core.components.collector import *
 from scheduler.core.components.ranker import RankerParameters, DefaultRanker
 from scheduler.core.eventsqueue.nightchanges import NightlyTimeline
 from scheduler.core.output import print_collector_info, print_plans
+from scheduler.core.plans import Plans
 from scheduler.core.eventsqueue import EveningTwilightEvent, EventQueue, MorningTwilightEvent, WeatherChangeEvent
+from scheduler.core.sources import Sources
 from scheduler.services import logger_factory
